requirements.txt
dotenv
streamlit
google.generativeai
pypdf
opencv-python-headless
python-docx
pillow